zmiany na mniejsze typy poziome bilbordy do zazbaczania

+++ src/BucketGrid.cpp
@@ -1,7 +1,7 @@
 #include "BucketGrid.h"
 #include "BucketIterator.h"
 
-BucketGrid::BucketGrid(int _resolution, double _size) {
+BucketGrid::BucketGrid(short _resolution, double _size) {
 	resolution = _resolution;
 	halfResolution = resolution / 2;
 	size = _size;
@@ -17,7 +17,7 @@ BucketGrid::BucketGrid(int _resolution, double _size) {
 		}
 	}
 
-	levelsCache = new std::vector<std::pair<int, int>*>*[RES_SEP_DIST];
+	levelsCache = new std::vector<std::pair<short, short>*>*[RES_SEP_DIST];
 	for (int i = 0; i < RES_SEP_DIST; ++i) {
 		levelsCache[i] = getEnvIndexs((((double)MAX_SEP_DIST) / RES_SEP_DIST) * i);
 	}
@@ -42,7 +42,7 @@ void BucketGrid::updateGrid(Entity* entity, int team) {
 	}
 }
 
-std::vector<std::pair<int, int>*>* BucketGrid::getEnvIndexsFromCache(double dist) {
+std::vector<std::pair<short, short>*>* BucketGrid::getEnvIndexsFromCache(double dist) {
 	int index = dist / diff;
 	return levelsCache[index];
 }
@@ -112,27 +112,27 @@ bool BucketGrid::fieldInCircle(int i, int j, double radius) {
 
 }
 
-std::vector<std::pair<int, int>*>* BucketGrid::getEnvIndexs(double radius) {
-	std::vector<std::pair<int, int>*>* indexes = new std::vector<std::pair<int, int>*>();
+std::vector<std::pair<short, short>*>* BucketGrid::getEnvIndexs(double radius) {
+	std::vector<std::pair<short, short>*>* indexes = new std::vector<std::pair<short, short>*>();
 	for (int i = 0; i < RES_SEP_DIST; i++) {
 		for (int j = 0; j < RES_SEP_DIST; j++) {
 			if (fieldInCircle(i, j, radius)) {
 				int x = i + 1;
 				int y = j + 1;
-				indexes->push_back(new std::pair<int, int>(x, y));
-				indexes->push_back(new std::pair<int, int>(x, -y));
-				indexes->push_back(new std::pair<int, int>(-x, y));
-				indexes->push_back(new std::pair<int, int>(-x, -y));
+				indexes->push_back(new std::pair<short, short>(x, y));
+				indexes->push_back(new std::pair<short, short>(x, -y));
+				indexes->push_back(new std::pair<short, short>(-x, y));
+				indexes->push_back(new std::pair<short, short>(-x, -y));
 			}
 		}
 		if (fieldInCircle(i, 0, radius)) {
 			int x = i + 1;
-			indexes->push_back(new std::pair<int, int>(x, 0));
-			indexes->push_back(new std::pair<int, int>(0, x));
-			indexes->push_back(new std::pair<int, int>(-x, 0));
-			indexes->push_back(new std::pair<int, int>(0, -x));
+			indexes->push_back(new std::pair<short, short>(x, 0));
+			indexes->push_back(new std::pair<short, short>(0, x));
+			indexes->push_back(new std::pair<short, short>(-x, 0));
+			indexes->push_back(new std::pair<short, short>(0, -x));
 		}
 	}
-	indexes->push_back(new std::pair<int, int>(0, 0));
+	indexes->push_back(new std::pair<short, short>(0, 0));
 	return indexes;
 }
