+++ src/BucketIterator.cpp
@@ -1,11 +1,11 @@
 #include "BucketIterator.h"
 #include "BucketGrid.h"
 
-BucketIterator::BucketIterator(std::vector<std::pair<int, int>*>* _levels, int _dX, int _dZ, BucketGrid* _bucketGrid) {
+BucketIterator::BucketIterator(std::vector<std::pair<short, short>*>* _levels, short _dX, short _dZ, BucketGrid* _bucketGrid) {
 	levels = _levels;
 	dX = _dX;
 	dZ = _dZ;
-	std::pair<int, int>* pair = (*levels)[index];
+	std::pair<short, short>* pair = (*levels)[index];
 	bucketGrid = _bucketGrid;
 	currentContent = bucketGrid->getBucketAt(pair->first + dX, pair->second + dZ)->getContent();
 	sizeContent = currentContent->size();
@@ -19,7 +19,7 @@ BucketIterator::~BucketIterator() {
 Entity* BucketIterator::next() {
 	while (secondIndex == sizeContent) {
 		if (index >= levelSize) { return nullptr; }
-		std::pair<int, int>* pair = (*levels)[index];
+		std::pair<short, short>* pair = (*levels)[index];
 		++index;
 		currentContent = bucketGrid->getBucketAt(pair->first + dX, pair->second + dZ)->getContent();
 		sizeContent = currentContent->size();
