+++ src/Gradient.h
@@ -7,7 +7,7 @@ class Entity;
 class Gradient
 {
 public:
-	Gradient(int _resolution, double _size);
+	Gradient(short _resolution, double _size);
 	~Gradient();
 	Urho3D::Vector3* getValueAt(double x, double z);
 	int getIntegerPos(double value);
@@ -18,8 +18,8 @@ public:
 	void remove(Entity * entity);
 private:
 	Urho3D::Vector3*** values;
-	int resolution;
-	int halfResolution;
+	short resolution;
+	short halfResolution;
 	double size;
 	double fieldSize;
 };
